+++ neuro_san/message/filters/chat_context_message_filter.py
@@ -17,13 +17,13 @@
 from typing import Any
 from typing import Dict
 
-from neuro_san.internals.filters.message_filter import MessageFilter
-from neuro_san.internals.messages.chat_message_type import ChatMessageType
+from neuro_san.message.filters.message_filter import MessageFilter
+from neuro_san.message.types.chat_message_type import ChatMessageType
 
 
-class MaximalMessageFilter(MessageFilter):
+class ChatContextMessageFilter(MessageFilter):
     """
-    MessageFilter implementation that lets everything through.
+    MessageFilter implementation for a message with "the chat_context" in it.
     """
 
     def allow_message(self, chat_message_dict: Dict[str, Any], message_type: ChatMessageType) -> bool:
@@ -34,8 +34,12 @@ class MaximalMessageFilter(MessageFilter):
         :param message_type: The ChatMessageType of the chat_message_dictionary to process.
         :return: True if the message should be allowed through to the client. False otherwise.
         """
-        # As long as the dictionary has some keys in it, we will pass it on.
-        if any(chat_message_dict):
-            return True
+        if message_type != ChatMessageType.AGENT_FRAMEWORK:
+            # Final chat_contexts are only ever AI Messages
+            return False
 
-        return False
+        if chat_message_dict.get("chat_context") is None:
+            return False
+
+        # Meets all our criteria. Let it through.
+        return True
