+++ neuro_san/message/filters/message_filter_factory.py
@@ -18,9 +18,9 @@ from typing import Any
 from typing import Dict
 from typing import Type
 
-from neuro_san.internals.filters.maximal_message_filter import MaximalMessageFilter
-from neuro_san.internals.filters.message_filter import MessageFilter
-from neuro_san.internals.filters.minimal_message_filter import MinimalMessageFilter
+from neuro_san.message.filters.maximal_message_filter import MaximalMessageFilter
+from neuro_san.message.filters.message_filter import MessageFilter
+from neuro_san.message.filters.minimal_message_filter import MinimalMessageFilter
 
 TYPE_TO_MESSAGE_FILTER_CLASS: Dict[Any, Type[MessageFilter]] = {
     0:  MinimalMessageFilter,
