+++ neuro_san/message/filters/token_accounting_message_filter.py
@@ -18,8 +18,8 @@ from typing import Any
 from typing import Dict
 from typing import List
 
-from neuro_san.internals.filters.message_filter import MessageFilter
-from neuro_san.internals.messages.chat_message_type import ChatMessageType
+from neuro_san.message.filters.message_filter import MessageFilter
+from neuro_san.message.types.chat_message_type import ChatMessageType
 
 
 class TokenAccountingMessageFilter(MessageFilter):
