+++ neuro_san/message/parsers/structure/structure_parser.py
