+++ neuro_san/message/processors/message_processor.py
@@ -18,7 +18,7 @@ from typing import Any
 from typing import Dict
 from typing import List
 
-from neuro_san.internals.messages.chat_message_type import ChatMessageType
+from neuro_san.message.types.chat_message_type import ChatMessageType
 
 
 class MessageProcessor:
