+++ neuro_san/message/types/agent_framework_message.py
@@ -29,8 +29,8 @@ from copy import copy
 
 from pydantic import ConfigDict
 
-from neuro_san.internals.messages.sly_data_redactor import SlyDataRedactor
-from neuro_san.internals.messages.traced_message import TracedMessage
+from neuro_san.message.types.traced_message import TracedMessage
+from neuro_san.message.utils.sly_data_redactor import SlyDataRedactor
 
 
 class AgentFrameworkMessage(TracedMessage):
